ActiveRecord config usable
