+++ app/indices/topic_index.rb
@@ -1,20 +1,5 @@
 # frozen_string_literal: true
 
-unless System::Database.oracle?
-  ThinkingSphinx::Index.define(:topic, with: :real_time) do
-    indexes :title
-    indexes sphinx_post_bodies, as: :post
-
-    has :tenant_id, type: :bigint
-
-    has :forum_id, type: :bigint
-    has :sticky, type: :boolean
-    has :last_updated_at, type: :timestamp
-
-    scope { Topic.includes(:posts) }
-  end
-end
-
 module TopicIndex
   extend ActiveSupport::Concern
 
@@ -59,5 +44,20 @@ module TopicIndex
   def allow_system_indexation?
     !System::Database.oracle? &&
       ThinkingSphinx::Configuration.new.settings['indexed_models'].include?('Topic')
+  end
+end
+
+unless System::Database.oracle?
+  ThinkingSphinx::Index.define(:topic, with: :real_time) do
+    indexes :title
+    indexes sphinx_post_bodies, as: :post
+
+    has :tenant_id, type: :bigint
+
+    has :forum_id, type: :bigint
+    has :sticky, type: :boolean
+    has :last_updated_at, type: :timestamp
+
+    scope { Topic.includes(:posts) }
   end
 end
