+++ config/initializers/active_record.rb
@@ -1,10 +1,14 @@
-# TODO: remove this hack?
-require 'deadlock_retry'
-class ActiveRecord::Base
+ActiveSupport.on_load(:active_record) do
 
-  # returns a human readable name of the class.
-  # It is here mainly to allow Cinstance to return "Application"
-  def human_name
-    self.to_s
+  # shall we still rely on this abandoned in 2012 retry code?
+  require 'deadlock_retry'
+
+  class ActiveRecord::Base
+
+    # returns a human readable name of the class.
+    # It is here mainly to allow Cinstance to return "Application"
+    def human_name
+      self.to_s
+    end
   end
 end
