+++ config/initializers/babel_squeel.rb
@@ -1,3 +1,5 @@
-BabySqueel.configure do |config|
-  config.enable_compatibility!
+ActiveSupport.on_load(:active_record) do
+  BabySqueel.configure do |config|
+    config.enable_compatibility!
+  end
 end
