+++ config/initializers/event_store.rb
@@ -1,5 +1,9 @@
-System::Application.configure do
-  config = self.config
+Rails.configuration.to_prepare do
+  require 'rails_event_store'
+  require 'three_scale/sidekiq_retry_support'
 
-  ActiveSupport::Reloader.to_prepare { config.event_store = EventStore::Repository.new }
+  System::Application.configure do
+    config = self.config
+    config.event_store = EventStore::Repository.new
+  end
 end
