+++ config/initializers/lazy_initialization.rb
@@ -1,4 +1,6 @@
 # frozen_string_literal: true
 
-require 'three_scale/lazy_initialization'
-ActiveRecord::Base.send(:include, ThreeScale::LazyInitialization)
+ActiveSupport.on_load(:active_record) do
+  require 'three_scale/lazy_initialization'
+  ActiveRecord::Base.send(:include, ThreeScale::LazyInitialization)
+end
