+++ config/initializers/money.rb
@@ -1,2 +1,4 @@
-ActiveRecord::Base.send(:include, ThreeScale::HasMoney) if defined?(ActiveRecord)
+ActiveSupport.on_load(:active_record) do
+  ActiveRecord::Base.send(:include, ThreeScale::HasMoney)
+end
 ActionController::Base.helper(ThreeScale::MoneyHelper) if defined?(ActionController)
