+++ config/initializers/ruby-2.rb
@@ -1,3 +1,7 @@
+ActiveSupport.on_load(:active_record) do
+
 ActiveRecord::Base.class_eval do
   public :initialize_dup
+end
+
 end
